# docs/api-spec/openapi.yml
openapi: 3.0.3
info:
  title: Authorization API
  description: 認可サーバーのAPI群です。
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: API開発チーム
    email: satoshi.nagashiba@gmail.com
servers:
  - url: https://apis.authorization.com
    description: 本番環境
  - url: https://apis-stg.authorization.com
    description: ステージング環境
  - url: https://apis-dev.authorization.com
    description: 開発環境
  - url: https://apis.authorization-php.dev/api
    description: ローカルPHP（Laravel）環境
  - url: https://apis.authorization-csharp.dev/api
    description: ローカルC#（ASP.NET Core）環境
  - url: https://apis.authorization-go-gin.dev/api
    description: ローカルGo（Gin）環境
  - url: https://apis.authorization-go-beego.dev/api
    description: ローカルGo（Beego）環境
  - url: https://apis.authorization-go-echo.dev/api
    description: ローカルGo（Echo）環境
  - url: https://apis.authorization-java.dev/api
    description: ローカルJava（Spring Boot）環境
  - url: https://apis.authorization-kotlin.dev/api
    description: ローカルKotlin（Ktor）環境
  - url: https://apis.authorization-python.dev/api
    description: ローカルPython（FastAPI）環境
  - url: https://apis.authorization-rb-hanami.dev/api
    description: ローカルRuby（Hanami）環境
  - url: https://apis.authorization-rb-rails.dev/api
    description: ローカルRuby（Rails）環境
  - url: https://apis.authorization-rust.dev/api
    description: ローカルRust（Axum）環境
  - url: https://apis.authorization-ts.dev/api
    description: ローカルTypeScript（Hono）環境
paths:
  /auth/login:
    get:
      operationId: getLogin
      summary: ログイン情報を取得します
      tags:
        - auth
      security: []
      responses:
        '200':
          description: 成功時にレスポンスを返します。
        '404':
          description: ログインページが存在しない場合のエラー
  /auth/invitation/{token}:
    get:
      operationId: getInvitation
      summary: 招待URLトークンを検証し、ログイン情報を取得します
      tags:
        - auth
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: 招待トークン
      responses:
        '200':
          description: 成功時にレスポンスを返します。
        '404':
          description: ログインページが存在しない場合のエラー
        '500':
          description: 招待トークンが不正な場合のエラー
  /auth/google/redirect:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: googleRedirect
      summary: Googleログインページへリダイレクトします
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        CSRF 対策として暗号学的に安全な nonce を生成し、HttpOnly クッキー `oauth_state`（有効期限 10 分）に保存したうえで
        state パラメータ（`{runtime}|{nonce}` または招待時 `{runtime}|{nonce}|{token}`）に埋め込む。
      tags:
        - auth
      security: []
      responses:
        '302':
          description: Google 認証ページへリダイレクト
        '400':
          description: リクエストが不正です
        '500':
          description: Googleログインに失敗した場合のエラー
  /auth/google/callback:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: googleCallback
      summary: Googleログインからのコールバックを受け取ります
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        state に含まれる nonce と `oauth_state` クッキーの値を定数時間比較で検証し、
        クッキー欠落・state 不正・nonce 不一致の場合は /error?code=400 へリダイレクトする。
        検証後は成否を問わず `oauth_state` クッキーを破棄する。
      tags:
        - auth
      security: []
      responses:
        '302':
          description: ログイン成功後フロントエンドへリダイレクト（state 検証失敗時は /error?code=400 へリダイレクト）
        '403':
          description: 招待フローで招待トークンが無効な場合
        '500':
          description: Googleログインに失敗した場合のエラー
  /auth/github/redirect:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: githubRedirect
      summary: GitHubログインページへリダイレクトします
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        state パラメータにバックエンド識別子（例: go-gin）を含め、
        GitHub からのコールバックを正しいバックエンドへ振り分けるために使用する。
        あわせて CSRF 対策として暗号学的に安全な nonce を生成し、HttpOnly クッキー `oauth_state`（有効期限 10 分）に保存したうえで
        state（`{runtime}|{nonce}` または招待時 `{runtime}|{nonce}|{token}`）に埋め込む。
      tags:
        - auth
      security: []
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
          description: 招待トークン（招待フロー経由の場合のみ）
      responses:
        '302':
          description: GitHub 認証ページへリダイレクト
        '400':
          description: リクエストが不正です
        '500':
          description: GitHubログインに失敗した場合のエラー
  /auth/github/callback:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: githubCallback
      summary: GitHubログインからのコールバックを受け取ります
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        GitHub は redirect_uri を1つしか登録できないため、フロントエンドの
        /auth/github/callback（Next.js ディスパッチャー）が state を読み取り、
        /function/{backend}/auth/github/callback へ転送してからこのエンドポイントに到達する。
        state に含まれる nonce と `oauth_state` クッキーの値を定数時間比較で検証し、
        クッキー欠落・state 不正・nonce 不一致の場合は /error?code=400 へリダイレクトする。
        検証後は成否を問わず `oauth_state` クッキーを破棄する。
      tags:
        - auth
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: GitHub から発行される認可コード
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: "`{runtime}|{nonce}` または `{runtime}|{nonce}|{token}`。先頭のバックエンド識別子（例: go-gin）をディスパッチャーがルーティングに使用し、nonce を CSRF 検証に使用する"
      responses:
        '302':
          description: ログイン成功後フロントエンドへリダイレクト（state 検証失敗時は /error?code=400 へリダイレクト）
        '403':
          description: 招待フローで招待トークンが無効な場合
        '500':
          description: GitHubログインに失敗した場合のエラー
  /auth/me:
    get:
      operationId: getMe
      summary: 自分自身のプロフィールを取得します
      description: staff_id クッキーで認証済みのユーザー情報を返します。ログイン判定・ヘッダー表示などに使用します。
      tags:
        - auth
      security: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '401':
          description: 未認証（staff_id クッキーなし、または復号失敗）
        '404':
          description: スタッフが存在しない
  /auth/logout:
    get:
      operationId: logout
      summary: ログアウトします
      tags:
        - auth
      security: []
      responses:
        '200':
          description: 成功時にレスポンスを返します。
        '401':
          description: 未認証
        '500':
          description: ログアウトに失敗した場合のエラー
  /clients:
    get:
      operationId: listClients
      summary: クライアント一覧
      tags:
        - clients
      security: []
      parameters:
        - name: keyword
          in: query
          required: false
          description: キーワード
          schema:
            type: string
        - name: start_from
          in: query
          required: false
          description: 利用開始日 From
          schema:
            type: string
            format: date
        - name: start_to
          in: query
          required: false
          description: 利用開始日 To
          schema:
            type: string
            format: date
        - name: statuses
          in: query
          required: false
          description: 状態（複数指定可）
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - name: limit
          in: query
          required: false
          description: 取得件数
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: page
          in: query
          required: false
          description: 現在ページ番号（フロントのページャー算出用）
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: sort
          in: query
          required: false
          description: ソート対象カラム
          schema:
            type: string
            enum: [name, status, created_at]
        - name: sort_type
          in: query
          required: false
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clients'
        '404':
          description: ユーザーが存在しない場合のエラー
  /clients/{id}:
    get:
      operationId: getClient
      summary: クライアント検索
      tags:
        - clients
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: クライアントID
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: ユーザーが存在しない場合のエラー
  /clients/store:
    post:
      operationId: createClient
      summary: クライアント登録
      tags:
        - clients
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "SUCCESS"
        '400':
          description: リクエストが不正です
        '500':
          description: サーバーエラー
  /clients/{identifier}/qr:
    get:
      operationId: getClientQr
      summary: スマホアプリ連携用QRコードデータを取得します
      description: クライアント登録通知メールのリンクを踏んだ際にフロントエンドから呼び出されます。QRコードに埋め込むディープリンクURLを返します。
      tags:
        - clients
      security: []
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
          description: クライアント識別子
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientQr'
        '404':
          description: クライアントが存在しない
  /clients/{identifier}/info:
    get:
      operationId: getClientInfo
      summary: スマホアプリ向けクライアント情報取得
      description: QRコードのディープリンク経由でスマホアプリが呼び出します。アクセストークンを含むクライアント情報を返します。
      tags:
        - clients
      security:
        - clientTokenAuth: []
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
          description: クライアント識別子
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientMobileInfo'
        '401':
          description: アクセストークンが不正
        '404':
          description: クライアントが存在しない
  /clients/{identifier}/start:
    patch:
      operationId: startClientByApp
      summary: スマホアプリからの利用開始
      description: |
        スマホアプリの操作で利用開始します。管理UIからの操作（PUT /clients/{id}/update）とは別エンドポイントです。
        Inactive/Suspended の場合は Active に遷移します。既に Active の場合もアクセストークンを返します。
        認証不要（identifier のみで呼び出し可能）。
      tags:
        - clients
      security: []
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
          description: クライアント識別子
      responses:
        '200':
          description: 利用開始成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "SUCCESS"
                  access_token:
                    type: string
                    description: アクセストークン
                    example: "a1b2c3d4e5f6..."
        '404':
          description: クライアントが存在しない
  /clients/{identifier}/stop:
    patch:
      operationId: stopClientByApp
      summary: スマホアプリからの利用停止
      description: スマホアプリの操作で利用停止します。管理UIからの操作（PUT /clients/{id}/update）とは別エンドポイントです。
      tags:
        - clients
      security:
        - clientTokenAuth: []
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
          description: クライアント識別子
      responses:
        '200':
          description: 利用停止成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "SUCCESS"
        '401':
          description: アクセストークンが不正
        '404':
          description: クライアントが存在しない
        '409':
          description: 既に利用停止済み
  /clients/{id}/update:
    put:
      operationId: updateClient
      summary: クライアント変更
      tags:
        - clients
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: クライアントID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdate'
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "SUCCESS"
        '400':
          description: リクエストが不正です
        '409':
          description: 楽観排他ロック競合（他のリクエストが先に更新済み）
        '500':
          description: サーバーエラー
  /clients/{id}/delete:
    delete:
      operationId: deleteClient
      summary: クライアント削除
      tags:
        - clients
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: クライアントID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - version
              properties:
                version:
                  type: integer
                  format: int64
                  description: 楽観排他ロック用バージョン番号
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "SUCCESS"
        '404':
          description: ユーザーが存在しない場合のエラー
        '409':
          description: 楽観排他ロック競合（他のリクエストが先に更新済み）
  /clients/{id}/jwt-histories:
    get:
      operationId: listClientJwtHistories
      summary: JWT履歴一覧
      tags:
        - clients
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: クライアントID
        - name: limit
          in: query
          required: false
          description: 取得件数
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: page
          in: query
          required: false
          description: 現在ページ番号（フロントのページャー算出用）
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: sort
          in: query
          required: false
          description: ソート対象カラム
          schema:
            type: string
            enum: [issue_at, member_id]
        - name: sort_type
          in: query
          required: false
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtHistories'
        '404':
          description: クライアントが存在しない場合のエラー
  /staffs:
    get:
      operationId: listStaffs
      summary: スタッフ一覧
      tags:
        - staff
      security: []
      parameters:
        - name: keyword
          in: query
          required: false
          description: キーワード
          schema:
            type: string
        - name: roles
          in: query
          required: false
          description: 権限（複数指定可）
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - name: statuses
          in: query
          required: false
          description: 状態（複数指定可）
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
        - name: limit
          in: query
          required: false
          description: 取得件数
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: page
          in: query
          required: false
          description: 現在ページ番号（フロントのページャー算出用）
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: sort
          in: query
          required: false
          description: ソート対象カラム
          schema:
            type: string
            enum: [name, role, status, created_at]
        - name: sort_type
          in: query
          required: false
          description: ソート順
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffList'
        '404':
          description: ユーザーが存在しない場合のエラー
  /staffs/{id}/updateRole:
    patch:
      operationId: updateStaffRole
      summary: 権限を更新します
      tags:
        - staff
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: スタッフID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
                - version
                - executor_id
              properties:
                role:
                  type: integer
                  format: int64
                  description: 権限種別（1=管理者、2=メンバー）
                version:
                  type: integer
                  format: int64
                  description: 楽観排他ロック用バージョン番号
                executor_id:
                  type: integer
                  format: int64
                  description: 操作を実行したスタッフID
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '400':
          description: リクエストが不正です
        '404':
          description: スタッフが存在しない場合のエラー
        '409':
          description: 楽観排他ロック競合（他のリクエストが先に更新済み）
        '500':
          description: サーバーエラー
  /staffs/{id}/restore:
    patch:
      operationId: restoreStaff
      summary: スタッフの論理削除を復元します
      tags:
        - staff
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: スタッフID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - executor_id
              properties:
                executor_id:
                  type: integer
                  format: int64
                  description: 操作を実行したスタッフID
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '404':
          description: スタッフが存在しない場合のエラー
  /staffs/{id}/delete:
    delete:
      operationId: deleteStaff
      summary: スタッフを論理削除します
      tags:
        - staff
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: スタッフID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - version
                - executor_id
              properties:
                version:
                  type: integer
                  format: int64
                  description: 楽観排他ロック用バージョン番号
                executor_id:
                  type: integer
                  format: int64
                  description: 操作を実行したスタッフID
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '404':
          description: スタッフが存在しない場合のエラー
        '409':
          description: 楽観排他ロック競合（他のリクエストが先に更新済み）
  /admin/invitation:
    get:
      operationId: getAdminInvitation
      summary: 招待URLを取得します
      tags:
        - admin
      security: []
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: 招待用トークン（ハッシュ）を含むURL
                    example: "7f3a9c2e1b8d4f6a0e5c7b2d9f1a4e8c0d3b6a9f2e5c8d1b4a7f0e3c6d9b2a5f8"
        '404':
          description: スタッフが存在しない場合のエラー
  /admin/invitation/issue:
    get:
      operationId: issueAdminInvitation
      summary: 招待URLを発行します
      tags:
        - admin
      security: []
      parameters:
        - name: role
          in: query
          required: false
          description: 招待するスタッフの権限（1=管理者, 2=メンバー）。省略時は 2（メンバー）。
          schema:
            type: integer
            enum: [1, 2]
            default: 2
          example: 2
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: 招待用トークン（ハッシュ）を含むURL
                    example: "/auth/invitation/7f3a9c2e1b8d4f6a0e5c7b2d9f1a4e8c0d3b6a9f2e5c8d1b4a7f0e3c6d9b2a5f8"
        '400':
          description: role パラメーターが不正
        '401':
          description: 未認証
        '500':
          description: サーバーエラー
  /gate/issue:
    get:
      operationId: issueGate
      summary: JWTを発行する
      tags:
        - gate
      description: クライアントに対してJWTを発行します。
      security:
        - bearerAuth: []    # アクセストークン
      parameters:
        - name: member
          in: query
          required: true
          schema:
            type: string
          description: クライアント会員ID
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "SUCCESS"
        '400':
          description: リクエストが不正です
        '500':
          description: サーバーエラー
  /gate/client/{identifier}/verify:
    get:
      operationId: verifyGate
      summary: JWTを検証する
      tags:
        - gate
      security:
        - bearerAuth: []
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
          description: クライアント識別名

        - name: token
          in: query
          required: true
          schema:
            type: string
          description: クライアント識別名
      responses:
        '200':
          description: 成功時にレスポンスを返します。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payload'
        '400':
          description: リクエストが不正です
        '404':
          description: JWTが見つかりませんでした。
        '500':
          description: サーバーエラー
...
  /notifications/counts:
    get:
      operationId: getNotificationCounts
      summary: 未読・全件などのサマリー
      description: |
        未読件数・全件数など、通知まわりの集計を返す。
        複数種類のカウントを一度に返す。
      tags:
        - notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 正常応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationCounts'
        '401':
          description: 未認証
        '500':
          description: サーバーエラー
...
components:
  schemas:
...
    NotificationCounts:
      type: object
      description: 複数種類のカウントを一度に返す
      additionalProperties:
        type: integer
        format: int64
      example:
        unread: 3
        total: 42
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    clientTokenAuth:
      type: http
      scheme: bearer
      description: クライアントのアクセストークン認証（JWT ではなく登録時に発行される固定トークン）
...